requirements.txt
requests
discord.py
beautifulsoup4
asyncio